No-op auth step definitions for first @wip
Authorization and authentication are irrelevant for the first
scenario.  Until we are developing a scenario that relies on an
authentication or authorization *failure*, we can assume that all
users of the app are authorized to upload.  In practice, this could
even be combined with basic auth at the proxy server level,
essentially creating a single authorized "account."