First failing step, and decision point
This step, visiting the order log upload page, is the first step where
we have to implement an interface for the user.  In a real development
situation, we could take several paths here - the immediate
instantiation of a framework is by no means necessary, and might be
overkill - however, we know already we're writing a Rails app, so
this is where we'll add Rails.