Add cucumber-rails and initialize schema.rb
We have to dump the schema.rb file, even though the schema is
currently empty, or Rails will complain when it tries to clone the
schema into the test database.  Cucumber is now running in the Rails
context and via a rake task, and we're back to the same failure we
had before we introduced Rails.