Add order_logs and new form with file field
We finally create our first model - the order log.  For now, it just
has a column for the source data (we'll copy the entire file into the
database) and the gross revenue; this is the only data that we'll need
in order to satisfy the current scenario that's in @wip.

The file upload doesn't do anything yet - this just satisfies the
failing step in the scenario of attaching a file to the form and
submitting it.  The create action currently just redirects to the new
action, to avoid missing template exceptions (since we'll never
actually have a 'order_logs/create' template).