Calculate & display gross revenue from order logs
This satisfies the last step of our first scenario.  We're calculating
the gross revenue by parsing the file and directly aggregating the
order costs for each line; no normalization of data has occurred yet.

There are no user-facing specs that require (or benefit from)
normalization, so in order to fulfill that requirement, we'll need
backend acceptance tests, or we'll need to build a user interface
for interacting with that data, and use more full-stack integration
tests (which are typically my preference).