Add users, authentication, and authorization
This is a big commit.  All scenarios in "user/uploads_file.feature"
are now passing.  We now have a concept of users who can provide 
authentication credentials to gain access to the system, but who
also need to be explicitly authorized to do anything but arrive at
our Limbo page.