Replace custom composer with money_column gem
No reason to re-invent the wheel on this one; there were several
additional features we wanted that money_column gives us without
continuing to hack our own solution.  Our wheel had served its
purpose, but as soon as someone asked for it to be round and
dynamic, we figured someone had already invented that.