Add simplecov to check test coverage
I'm not, in this app, going to put as much effort into getting 100%
coverage, mostly because of the OpenID aspect - testing OpenID
properly without a bunch of tautological stubbing would take a lot
more time.  My approach would be to create a simple rack server that
simulates an OpenID server, and run that in the background whenever
running full stack tests that test authentication.  For now, I'm just
going to tackle some low-hanging fruit, refactor a bit to isolate
the difficult-to-test code, and move on.