Add letter_opener_web for dev environment email
This allows you to view emails sent in the development environment
in a web interface (mounted at /letter_opener), instead of having to
set up mail delivery or parse the logs.  This is just a web interface
to letter_opener (and it also doesn't launch a browser for every
email sent.. ugh).